Refactor exit command
