+++ data/brownie/tasklist.txt
@@ -1,5 +1,1 @@
-T | 0 | Complete CS2103 IP
-D | 0 | Finish sewing a pillowcase | 2025-02-10T10:00
 T | 0 | something
-E | 0 | something | 2020-01-01T10:00 | 2020-01-01T10:00
-D | 0 | something | 2020-01-01T10:00
